refactor: lottonumber -> lottoticket 패키지명 수정
